refactor(leaderboard): extract fetchLeaderboardData into shared lib
Pulls the data-assembly logic out of the cookie-authed /api/leaderboard
route so the upcoming /api/leaderboard-image route can reuse it without
duplicating the pipeline/targets/score math.
