+++ src/features/leaderboard/lib/fetch-leaderboard.ts
@@ -1,0 +1,375 @@
+import prisma from "@/lib/prisma";
+import { calculateTier, calculateCombinedScore } from "@/features/leaderboard/lib/scoring";
+import { getRepActuals } from "@/lib/opportunity-actuals";
+import type { LeaderboardEntry } from "@/features/leaderboard/lib/types";
+
+export interface LeaderboardTeamTotals {
+  revenue: number;
+  revenueCurrentFY: number;
+  revenuePriorFY: number;
+  unassignedRevenue: number;
+  unassignedRevenueCurrentFY: number;
+  unassignedRevenuePriorFY: number;
+  priorYearRevenue: number;
+  minPurchasesCurrentFY: number;
+  minPurchasesPriorFY: number;
+  unassignedPriorYearRevenue: number;
+  unassignedMinPurchasesCurrentFY: number;
+  unassignedMinPurchasesPriorFY: number;
+  pipelineCurrentFY: number;
+  pipelineNextFY: number;
+  unassignedPipelineCurrentFY: number;
+  unassignedPipelineNextFY: number;
+  targetedCurrentFY: number;
+  targetedNextFY: number;
+  unassignedTargetedCurrentFY: number;
+  unassignedTargetedNextFY: number;
+}
+
+export interface LeaderboardPayload {
+  initiative: {
+    id: number;
+    name: string;
+    startDate: string;
+    endDate: string | null;
+    showName: boolean;
+    showDates: boolean;
+    initiativeWeight: number;
+    pipelineWeight: number;
+    takeWeight: number;
+    revenueWeight: number;
+    revenueTargetedWeight: number;
+    pipelineFiscalYear: string | null;
+    takeFiscalYear: string | null;
+    revenueFiscalYear: string | null;
+    revenueTargetedFiscalYear: string | null;
+  };
+  fiscalYears: { currentFY: string; nextFY: string; priorFY: string };
+  entries: LeaderboardEntry[];
+  teamTotals: LeaderboardTeamTotals;
+  metrics: { action: string; label: string; pointValue: number }[];
+  thresholds: { tier: string; minPoints: number }[];
+}
+
+export class NoActiveInitiativeError extends Error {
+  constructor() { super("No active initiative"); this.name = "NoActiveInitiativeError"; }
+}
+
+/**
+ * Fetches the full leaderboard payload for the active initiative.
+ * Used by both /api/leaderboard (cookie-authed) and /api/leaderboard-image
+ * (bearer-secret-authed). Throws NoActiveInitiativeError if no active
+ * initiative exists.
+ */
+export async function fetchLeaderboardData(): Promise<LeaderboardPayload> {
+  const initiative = await prisma.initiative.findFirst({
+    where: { isActive: true },
+    include: { thresholds: true, metrics: true },
+  });
+
+  if (!initiative) {
+    throw new NoActiveInitiativeError();
+  }
+
+  const scores = await prisma.initiativeScore.findMany({
+    where: { initiativeId: initiative.id },
+    include: {
+      user: { select: { id: true, fullName: true, avatarUrl: true, email: true, role: true } },
+    },
+    orderBy: { totalPoints: "desc" },
+  });
+
+  const now = new Date();
+  const currentFY = now.getMonth() >= 6 ? now.getFullYear() + 1 : now.getFullYear();
+  const defaultSchoolYr = `${currentFY - 1}-${String(currentFY).slice(-2)}`;
+  const priorFY = currentFY - 1;
+  const priorSchoolYr = `${priorFY - 1}-${String(priorFY).slice(-2)}`;
+  const nextFYSchoolYr = `${currentFY}-${String(currentFY + 1).slice(-2)}`;
+
+  const pipelineSchoolYr = initiative.pipelineFiscalYear ?? defaultSchoolYr;
+  const takeSchoolYr = initiative.takeFiscalYear ?? defaultSchoolYr;
+  const revenueSchoolYr = initiative.revenueFiscalYear ?? defaultSchoolYr;
+
+  const uniqueYears = [...new Set([pipelineSchoolYr, takeSchoolYr, revenueSchoolYr, priorSchoolYr, defaultSchoolYr, nextFYSchoolYr])];
+
+  const repActuals = await Promise.all(
+    scores.map(async (score) => {
+      const email = score.user.email;
+      try {
+        const yearActuals = new Map<string, Awaited<ReturnType<typeof getRepActuals>>>();
+        await Promise.all(
+          uniqueYears.map(async (yr) => {
+            const actuals = await getRepActuals(email, yr);
+            yearActuals.set(yr, actuals);
+          })
+        );
+        return {
+          userId: score.userId,
+          pipeline: yearActuals.get(pipelineSchoolYr)?.openPipeline ?? 0,
+          pipelineCurrentFY: yearActuals.get(defaultSchoolYr)?.openPipeline ?? 0,
+          pipelineNextFY: yearActuals.get(nextFYSchoolYr)?.openPipeline ?? 0,
+          take: yearActuals.get(takeSchoolYr)?.totalTake ?? 0,
+          revenue: yearActuals.get(revenueSchoolYr)?.totalRevenue ?? 0,
+          revenueCurrentFY: yearActuals.get(defaultSchoolYr)?.totalRevenue ?? 0,
+          revenuePriorFY: yearActuals.get(priorSchoolYr)?.totalRevenue ?? 0,
+          priorYearRevenue: yearActuals.get(priorSchoolYr)?.minPurchaseBookings ?? 0,
+          minPurchasesCurrentFY: yearActuals.get(defaultSchoolYr)?.minPurchaseBookings ?? 0,
+          minPurchasesPriorFY: yearActuals.get(priorSchoolYr)?.minPurchaseBookings ?? 0,
+        };
+      } catch {
+        return {
+          userId: score.userId,
+          take: 0, pipeline: 0, pipelineCurrentFY: 0, pipelineNextFY: 0,
+          revenue: 0, revenueCurrentFY: 0, revenuePriorFY: 0,
+          priorYearRevenue: 0, minPurchasesCurrentFY: 0, minPurchasesPriorFY: 0,
+        };
+      }
+    })
+  );
+
+  const actualsMap = new Map(repActuals.map((a) => [a.userId, a]));
+  const adminUserIds = new Set(scores.filter((s) => s.user.role === "admin").map((s) => s.userId));
+  const rosterScores = scores.filter((s) => !adminUserIds.has(s.userId));
+  const rosterActuals = repActuals.filter((a) => !adminUserIds.has(a.userId));
+
+  const maxInitiativePoints = Math.max(...rosterScores.map((s) => s.totalPoints), 0);
+  const maxTake = Math.max(...rosterActuals.map((a) => a.take), 0);
+  const maxPipeline = Math.max(...rosterActuals.map((a) => a.pipeline), 0);
+  const maxRevenue = Math.max(...rosterActuals.map((a) => a.revenue), 0);
+
+  const thresholdData = initiative.thresholds.map((t) => ({ tier: t.tier, minPoints: t.minPoints }));
+
+  const userIds = scores.map((s) => s.userId);
+  const sinceDate = initiative.startDate;
+
+  const allPlans = await prisma.territoryPlan.findMany({
+    where: {
+      createdAt: { gte: sinceDate },
+      OR: [{ ownerId: { in: userIds } }, { userId: { in: userIds }, ownerId: null }],
+    },
+    select: { id: true, ownerId: true, userId: true },
+  });
+
+  const planCountMap = new Map<string, number>();
+  for (const plan of allPlans) {
+    const uid = plan.ownerId ?? plan.userId;
+    if (!uid) continue;
+    planCountMap.set(uid, (planCountMap.get(uid) ?? 0) + 1);
+  }
+
+  const planIds = allPlans.map((p) => p.id);
+
+  const [activityCounts, planDistricts] = await Promise.all([
+    prisma.activity.groupBy({
+      by: ["createdByUserId"],
+      where: { createdByUserId: { in: userIds }, createdAt: { gte: sinceDate } },
+      _count: true,
+    }),
+    prisma.territoryPlanDistrict.findMany({
+      where: { planId: { in: planIds } },
+      select: {
+        renewalTarget: true, winbackTarget: true, expansionTarget: true, newBusinessTarget: true,
+        plan: { select: { ownerId: true, userId: true } },
+      },
+    }),
+  ]);
+
+  const activityCountMap = new Map(activityCounts.map((a) => [a.createdByUserId, a._count]));
+
+  const revenueByUser = new Map<string, number>();
+  for (const d of planDistricts) {
+    const uid = d.plan.ownerId ?? d.plan.userId;
+    if (!uid) continue;
+    const total = Number(d.renewalTarget ?? 0) + Number(d.winbackTarget ?? 0) +
+                  Number(d.expansionTarget ?? 0) + Number(d.newBusinessTarget ?? 0);
+    revenueByUser.set(uid, (revenueByUser.get(uid) ?? 0) + total);
+  }
+
+  const currentFYInt = currentFY;
+  const nextFYInt = currentFY + 1;
+
+  const ownerFilter = {
+    OR: [{ ownerId: { in: userIds } }, { userId: { in: userIds }, ownerId: null }],
+  };
+
+  const emailByUserId = new Map<string, string>();
+  for (const s of scores) { if (s.user.email) emailByUserId.set(s.userId, s.user.email); }
+  const rosterEmails = [...emailByUserId.values()];
+
+  const [targetedCurrentFYDistricts, targetedNextFYDistricts, pipelineRows] = await Promise.all([
+    prisma.territoryPlanDistrict.findMany({
+      where: { plan: { ...ownerFilter, fiscalYear: currentFYInt } },
+      select: {
+        districtLeaid: true,
+        renewalTarget: true, winbackTarget: true, expansionTarget: true, newBusinessTarget: true,
+        plan: { select: { ownerId: true, userId: true } },
+      },
+    }),
+    prisma.territoryPlanDistrict.findMany({
+      where: { plan: { ...ownerFilter, fiscalYear: nextFYInt } },
+      select: {
+        districtLeaid: true,
+        renewalTarget: true, winbackTarget: true, expansionTarget: true, newBusinessTarget: true,
+        plan: { select: { ownerId: true, userId: true } },
+      },
+    }),
+    rosterEmails.length === 0
+      ? Promise.resolve([])
+      : prisma.$queryRaw<{ sales_rep_email: string; district_lea_id: string; school_yr: string; pipeline: number }[]>`
+          SELECT sales_rep_email, district_lea_id, school_yr,
+                 SUM(open_pipeline)::float AS pipeline
+          FROM district_opportunity_actuals
+          WHERE sales_rep_email = ANY(${rosterEmails})
+            AND school_yr IN (${defaultSchoolYr}, ${nextFYSchoolYr})
+          GROUP BY sales_rep_email, district_lea_id, school_yr
+          HAVING SUM(open_pipeline) > 0
+        `,
+  ]);
+
+  const repPipelineMap = new Map<string, number>();
+  for (const row of pipelineRows) {
+    repPipelineMap.set(`${row.sales_rep_email}::${row.district_lea_id}::${row.school_yr}`, Number(row.pipeline));
+  }
+
+  function sumTargetsWithPipelineDeduction(
+    districts: typeof targetedCurrentFYDistricts,
+    schoolYr: string,
+  ): Map<string, number> {
+    const map = new Map<string, number>();
+    for (const d of districts) {
+      const uid = d.plan.ownerId ?? d.plan.userId;
+      if (!uid) continue;
+      const target = Number(d.renewalTarget ?? 0) + Number(d.winbackTarget ?? 0) +
+                     Number(d.expansionTarget ?? 0) + Number(d.newBusinessTarget ?? 0);
+      if (target <= 0) continue;
+      const email = emailByUserId.get(uid);
+      const pipeline = email ? repPipelineMap.get(`${email}::${d.districtLeaid}::${schoolYr}`) ?? 0 : 0;
+      map.set(uid, (map.get(uid) ?? 0) + Math.max(0, target - pipeline));
+    }
+    return map;
+  }
+
+  const targetedCurrentFYByUser = sumTargetsWithPipelineDeduction(targetedCurrentFYDistricts, defaultSchoolYr);
+  const targetedNextFYByUser = sumTargetsWithPipelineDeduction(targetedNextFYDistricts, nextFYSchoolYr);
+
+  const revenueTargetedByUser = new Map<string, number>();
+  for (const uid of userIds) {
+    revenueTargetedByUser.set(uid, (targetedCurrentFYByUser.get(uid) ?? 0) + (targetedNextFYByUser.get(uid) ?? 0));
+  }
+  const maxRevenueTargeted = Math.max(...rosterScores.map((s) => revenueTargetedByUser.get(s.userId) ?? 0), 0);
+
+  const getActionCount = (userId: string, action: string): number => {
+    if (action === "plan_created") return planCountMap.get(userId) ?? 0;
+    if (action === "activity_logged") return activityCountMap.get(userId) ?? 0;
+    if (action === "revenue_targeted") return Math.floor((revenueByUser.get(userId) ?? 0) / 10000);
+    return 0;
+  };
+
+  const entries: LeaderboardEntry[] = rosterScores.map((score, index) => {
+    const actuals = actualsMap.get(score.userId) ?? {
+      userId: score.userId, take: 0, pipeline: 0, pipelineCurrentFY: 0, pipelineNextFY: 0,
+      revenue: 0, revenueCurrentFY: 0, revenuePriorFY: 0,
+      priorYearRevenue: 0, minPurchasesCurrentFY: 0, minPurchasesPriorFY: 0,
+    };
+    const tier = calculateTier(score.totalPoints, thresholdData);
+    const combinedScore = calculateCombinedScore({
+      initiativePoints: score.totalPoints, maxInitiativePoints,
+      pipeline: actuals.pipeline, maxPipeline,
+      take: actuals.take, maxTake,
+      revenue: actuals.revenue, maxRevenue,
+      revenueTargeted: revenueTargetedByUser.get(score.userId) ?? 0, maxRevenueTargeted,
+      initiativeWeight: Number(initiative.initiativeWeight),
+      pipelineWeight: Number(initiative.pipelineWeight),
+      takeWeight: Number(initiative.takeWeight),
+      revenueWeight: Number(initiative.revenueWeight),
+      revenueTargetedWeight: Number(initiative.revenueTargetedWeight),
+    });
+    const initiativeScore = maxInitiativePoints > 0 ? (score.totalPoints / maxInitiativePoints) * 100 : 0;
+    const pointBreakdown = initiative.metrics.map((m) => {
+      const count = getActionCount(score.userId, m.action);
+      return { action: m.action, label: m.label, pointValue: m.pointValue, count, total: count * m.pointValue };
+    });
+    return {
+      userId: score.userId,
+      fullName: score.user.fullName ?? "Unknown",
+      avatarUrl: score.user.avatarUrl,
+      totalPoints: score.totalPoints,
+      tier, rank: index + 1,
+      take: actuals.take, pipeline: actuals.pipeline,
+      pipelineCurrentFY: actuals.pipelineCurrentFY, pipelineNextFY: actuals.pipelineNextFY,
+      revenue: actuals.revenue, revenueCurrentFY: actuals.revenueCurrentFY, revenuePriorFY: actuals.revenuePriorFY,
+      priorYearRevenue: actuals.priorYearRevenue,
+      minPurchasesCurrentFY: actuals.minPurchasesCurrentFY, minPurchasesPriorFY: actuals.minPurchasesPriorFY,
+      revenueTargeted: revenueTargetedByUser.get(score.userId) ?? 0,
+      targetedCurrentFY: targetedCurrentFYByUser.get(score.userId) ?? 0,
+      targetedNextFY: targetedNextFYByUser.get(score.userId) ?? 0,
+      combinedScore: Math.round(combinedScore * 10) / 10,
+      initiativeScore: Math.round(initiativeScore * 10) / 10,
+      pointBreakdown,
+    };
+  });
+
+  const sumActuals = (
+    pool: typeof repActuals,
+    key:
+      | "revenue" | "revenueCurrentFY" | "revenuePriorFY"
+      | "priorYearRevenue" | "minPurchasesCurrentFY" | "minPurchasesPriorFY"
+      | "pipelineCurrentFY" | "pipelineNextFY",
+  ): number => pool.reduce((acc, a) => acc + (a[key] ?? 0), 0);
+
+  const sumTargetedMap = (pool: Map<string, number>, ids: Iterable<string>): number => {
+    let total = 0;
+    for (const id of ids) total += pool.get(id) ?? 0;
+    return total;
+  };
+
+  const adminActuals = repActuals.filter((a) => adminUserIds.has(a.userId));
+
+  const teamTotals: LeaderboardTeamTotals = {
+    revenue: sumActuals(repActuals, "revenue"),
+    revenueCurrentFY: sumActuals(repActuals, "revenueCurrentFY"),
+    revenuePriorFY: sumActuals(repActuals, "revenuePriorFY"),
+    unassignedRevenue: sumActuals(adminActuals, "revenue"),
+    unassignedRevenueCurrentFY: sumActuals(adminActuals, "revenueCurrentFY"),
+    unassignedRevenuePriorFY: sumActuals(adminActuals, "revenuePriorFY"),
+    priorYearRevenue: sumActuals(repActuals, "priorYearRevenue"),
+    minPurchasesCurrentFY: sumActuals(repActuals, "minPurchasesCurrentFY"),
+    minPurchasesPriorFY: sumActuals(repActuals, "minPurchasesPriorFY"),
+    unassignedPriorYearRevenue: sumActuals(adminActuals, "priorYearRevenue"),
+    unassignedMinPurchasesCurrentFY: sumActuals(adminActuals, "minPurchasesCurrentFY"),
+    unassignedMinPurchasesPriorFY: sumActuals(adminActuals, "minPurchasesPriorFY"),
+    pipelineCurrentFY: sumActuals(repActuals, "pipelineCurrentFY"),
+    pipelineNextFY: sumActuals(repActuals, "pipelineNextFY"),
+    unassignedPipelineCurrentFY: sumActuals(adminActuals, "pipelineCurrentFY"),
+    unassignedPipelineNextFY: sumActuals(adminActuals, "pipelineNextFY"),
+    targetedCurrentFY: sumTargetedMap(targetedCurrentFYByUser, userIds),
+    targetedNextFY: sumTargetedMap(targetedNextFYByUser, userIds),
+    unassignedTargetedCurrentFY: sumTargetedMap(targetedCurrentFYByUser, adminUserIds),
+    unassignedTargetedNextFY: sumTargetedMap(targetedNextFYByUser, adminUserIds),
+  };
+
+  return {
+    initiative: {
+      id: initiative.id,
+      name: initiative.name,
+      startDate: initiative.startDate.toISOString(),
+      endDate: initiative.endDate?.toISOString() ?? null,
+      showName: initiative.showName,
+      showDates: initiative.showDates,
+      initiativeWeight: Number(initiative.initiativeWeight),
+      pipelineWeight: Number(initiative.pipelineWeight),
+      takeWeight: Number(initiative.takeWeight),
+      revenueWeight: Number(initiative.revenueWeight),
+      revenueTargetedWeight: Number(initiative.revenueTargetedWeight),
+      pipelineFiscalYear: initiative.pipelineFiscalYear,
+      takeFiscalYear: initiative.takeFiscalYear,
+      revenueFiscalYear: initiative.revenueFiscalYear,
+      revenueTargetedFiscalYear: initiative.revenueTargetedFiscalYear,
+    },
+    fiscalYears: { currentFY: defaultSchoolYr, nextFY: nextFYSchoolYr, priorFY: priorSchoolYr },
+    entries,
+    teamTotals,
+    metrics: initiative.metrics.map((m) => ({ action: m.action, label: m.label, pointValue: m.pointValue })),
+    thresholds: initiative.thresholds.map((t) => ({ tier: t.tier, minPoints: t.minPoints })),
+  };
+}
